requirements.txt
# requirements.txt
streamlit
google-cloud-speech
google-cloud-texttospeech
openai
moviepy